Move remaining match components to dedicated folder
