+++ osu.Game/Screens/Multi/Match/Components/MatchSongSelect.cs
@@ -10,10 +10,9 @@ using osu.Framework.Screens;
 using osu.Game.Beatmaps;
 using osu.Game.Online.Multiplayer;
 using osu.Game.Rulesets.Mods;
-using osu.Game.Screens.Multi;
-using osu.Game.Screens.Play;
+using osu.Game.Screens.Select;
 
-namespace osu.Game.Screens.Select
+namespace osu.Game.Screens.Multi.Match.Components
 {
     public class MatchSongSelect : SongSelect, IMultiplayerSubScreen
     {
